73-FE: Update finance record composables to inject search filters
Soon, we'll need to invalidate multiple query keys, so passing a single
query key won't suffice. The composables will need direct access to
the search filters.

Since these composables will only be used on the finances page, we can
reduce testing a little by injecting rather than passing props (If
passing props, a test should be added for the caller that checks that
the caller provides filters to the composables).