120-FE: Update `useDeleteFinanceRecord` to invalidate stats instead of messing with the query cache
With the introduction of stats by date, manually updating the cached
stats is a lot more complicated. Although it's possible to figure out
which date "bucket" the deleted record belonged to and update the
appropriate stats fields, I don't think the time & effort needed to
implement and maintain such an approach is worth it.